requirements.txt
torch==v1.12.0
torchtext==v0.13.0
pyyaml==v6.0
numpy==v1.26.4
